Moved sortTemplate and added sorting of function namespaces
